Improve layouts of Youtube Search
With plans to release this application I decided to improve the scaling of the layout. I changed the Retrofit's call response to a higher resolution of the thumbnail. Next I scaled it properly to fit well to multiple devices. Also removed redundant initialization of FragmentManager object in the Search_Adapter