Add Update method to Room's Dao and Fix the Deletesong not Showing the Songs
I implemented the update function for my room's dao as it seems to finally be working with androidx. It removed the need to remove and then add playlist hence reducing confusing code. I also fixed the Observer issue in the DeleteSong fragment so now its showing and deleting the songs correctly. Still to do:
-Use Rxjava for background thread queries in Room
-Ensure no other fragments insances are made when commiting fragments.
-Look into service class